+++ Cargo.toml
@@ -11,6 +11,8 @@ galvan-files = { path = "galvan-files", version = "0.0.0-dev09" }
 galvan-pest = { path = "galvan-pest", version = "0.0.0-dev09" }
 galvan-test-macro = { path = "galvan-test-macro", version = "0.0.0-dev09" }
 galvan-resolver = { path = "galvan-resolver", version = "0.0.0-dev09" }
+galvan-into-ast = { path = "galvan-into-ast", version = "0.0.0-dev09" }
+galvan-parse = { path = "galvan-parse", version = "0.0.0-dev09" }
 
 thiserror = "1.0.51"
 itertools = "0.12.0"
@@ -38,6 +40,8 @@ build = ["galvan-transpiler/exec"]
 resolver = "2"
 members = [
   "galvan-ast",
+  "galvan-into-ast",
+  "galvan-parse",
   # "galvan-core",
   "galvan-files",
   "galvan-pest",
@@ -45,5 +49,5 @@ members = [
   "galvan-transpiler",
   "galvan-test-macro",
   "galvan-test",
-  "tree-sitter-galvan", "galvan-into-ast",
+  "tree-sitter-galvan",
 ]
