+++ galvan-ast/src/lib.rs
@@ -9,14 +9,23 @@ mod item;
 
 pub use item::*;
 
-mod result;
-use result::Result;
-pub use result::{AstError, AstResult};
-
 #[derive(Debug, PartialEq, Eq)]
 pub struct Ast {
     pub toplevel: Vec<RootItem>,
     pub source: Source,
+}
+
+impl Ast {
+    pub fn with_source(self, source: Source) -> Ast {
+        if self.source != Source::Missing {
+            panic!("Attempting to set a source to an AST that already had a source!");
+        }
+
+        Ast {
+            toplevel: self.toplevel,
+            source,
+        }
+    }
 }
 
 #[derive(Debug, PartialEq, Eq)]
@@ -46,80 +55,4 @@ pub struct SegmentedAsts {
     // pub other: Vec<ToplevelItem<CustomTaskDecl>>
 }
 
-pub trait SegmentAst {
-    fn segmented(self) -> Result<SegmentedAsts>;
-}
 
-impl SegmentAst for Ast {
-    fn segmented(self) -> Result<SegmentedAsts> {
-        let mut types = Vec::new();
-        let mut functions = Vec::new();
-        let mut tests = Vec::new();
-        let mut main = None;
-
-        for item in self.toplevel {
-            match item {
-                RootItem::Type(item) => types.push(ToplevelItem {
-                    item,
-                    source: self.source.clone(),
-                }),
-                RootItem::Fn(item) => functions.push(ToplevelItem {
-                    item,
-                    source: self.source.clone(),
-                }),
-                RootItem::Test(item) => tests.push(ToplevelItem {
-                    item,
-                    source: self.source.clone(),
-                }),
-                RootItem::Main(item) => {
-                    if main.is_some() {
-                        return Err(AstError::DuplicateMain);
-                    }
-
-                    main = Some(ToplevelItem {
-                        item,
-                        source: self.source.clone(),
-                    })
-                }
-            }
-        }
-
-        Ok(SegmentedAsts {
-            types,
-            functions,
-            tests,
-            main,
-        })
-    }
-}
-
-impl SegmentAst for Vec<Ast> {
-    fn segmented(self) -> Result<SegmentedAsts> {
-        let mut types = Vec::new();
-        let mut functions = Vec::new();
-        let mut tests = Vec::new();
-        let mut main = None;
-        let segmented = self.into_iter().map(SegmentAst::segmented);
-
-        for ast in segmented {
-            let ast = ast?;
-            types.extend(ast.types);
-            functions.extend(ast.functions);
-            tests.extend(ast.tests);
-            if let Some(main_decl) = ast.main {
-                if main.is_some() {
-                    return Err(AstError::DuplicateMain);
-                }
-
-                main = Some(main_decl);
-            }
-        }
-
-        Ok(SegmentedAsts {
-            types,
-            functions,
-            tests,
-            main,
-        })
-    }
-}
