+++ galvan-parse/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "galvan-parse"
+authors.workspace = true
+edition.workspace = true
+version.workspace = true
+license.workspace = true
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+tree-sitter-galvan = { path = "../tree-sitter-galvan" } 
+
+galvan-files.workspace = true
